+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
   * Browser.assert.cookie
   * Cookies submodule
+  * Remove default assertions errors
 
 0.4.5 / 2018-03-24
 ==================